fix: find_unused_parameters with SageMaker DDP
HF Trainer (as of at least v4.17 to 4.19.2) ignores the configuration
of `ddp_find_unused_parameters` when running in SageMaker Distributed
Data Parallel (vs PyTorch native DDP). LayoutLMv2 requires this
setting to train in DDP mode, and from initial testing SMDDP seems to
support the parameter no problem - so set this param as true by
default and patch the Trainer to support passing it through.