feat(nbs): correct total time on textract pstart
Mark the "Starting jobs..." progress as completed, and display the
correct elapsed time, as soon as all jobs have been started by
call_textract(). Previously the job start bar was updated, but left
open: So the total run time was later updated to the *overall* time
including waiting for jobs to complete.